Fix some uses of `cpointer` that should be `cstring`.
Prior to this commit it was possible to run into
issues with invalid characters in operation names
because these Savi `String`s were being sometimes
allocated without a null terminator (which is
normal and okay, until you try to pass it into
a C function that is expecting a null-terminated string.

The right fix is to use `cstring` instead of `cpointer`
in these cases.